🤖 Move BasicAuthHandlingImplConfig to models; add metricsforwarder config parser
Move the sealed interface BasicAuthHandlingImplConfig and its
implementations (BasicAuthHandlingNative, BasicAuthHandlingStoredProc)
from api/config to the models package so both api/config and
metricsforwarder/config can use them without circular imports.

Introduce metricsforwarder/config/parser.go with rawConfig/toConfig
pattern (analog to api/config/parser.go). The metricsforwarder Config
uses its own CredentialHelperConfig field (models.BasicAuthHandlingImplConfig)
instead of reusing the apiserver CustomMetricsAuthConfig.
