+++ api/broker/broker_suite_test.go
@@ -54,7 +54,7 @@ var _ = BeforeSuite(func() {
 		CustomMetricsAuthConfig: &config.CustomMetricsAuthConfig{
 			BasicAuthHandling: config.BasicAuthHandlingOn,
 			DefaultCustomMetricAuthType: models.BindingSecret,
-			BasicAuthHandlingImplConfig: config.BasicAuthHandlingNative{},
+			BasicAuthHandlingImplConfig: models.BasicAuthHandlingNative{},
 		},
 	}
