+++ api/cmd/api/api_suite_test.go
@@ -209,7 +209,7 @@ var _ = SynchronizedBeforeSuite(func() []byte {
 	conf.CustomMetricsAuthConfig = &config.CustomMetricsAuthConfig{
 		BasicAuthHandling:           config.BasicAuthHandlingOn,
 		DefaultCustomMetricAuthType: models.BindingSecret,
-		BasicAuthHandlingImplConfig: config.BasicAuthHandlingNative{},
+		BasicAuthHandlingImplConfig: models.BasicAuthHandlingNative{},
 	}
 
 	configFile = writeConfig(&conf)
