+++ cred_helper/custom_metrics_cred_helper.go
@@ -8,7 +8,6 @@ import (
 	"os"
 	"time"
 
-	"code.cloudfoundry.org/app-autoscaler/src/autoscaler/api/config"
 	"code.cloudfoundry.org/app-autoscaler/src/autoscaler/db/sqldb"
 	"github.com/google/uuid"
 
@@ -37,10 +36,10 @@ func (c *customMetricsCredentials) Close() error {
 	return c.policyDB.Close()
 }
 
-func CredentialsProvider(implConfig config.BasicAuthHandlingImplConfig, dbConf map[string]db.DatabaseConfig, cacheTTL time.Duration, cacheCleanupInterval time.Duration, logger lager.Logger, policyDB db.PolicyDB) Credentials {
+func CredentialsProvider(implConfig models.BasicAuthHandlingImplConfig, dbConf map[string]db.DatabaseConfig, cacheTTL time.Duration, cacheCleanupInterval time.Duration, logger lager.Logger, policyDB db.PolicyDB) Credentials {
 	var credentials Credentials
 	switch impl := implConfig.(type) {
-	case config.BasicAuthHandlingStoredProc:
+	case models.BasicAuthHandlingStoredProc:
 		storedProcedureDb, err := sqldb.NewStoredProcedureSQLDb(impl.Config, dbConf[db.StoredProcedureDb], logger.Session("storedprocedure-db"))
 		if err != nil {
 			logger.Fatal("failed to connect to storedProcedureDb database", err, lager.Data{"dbConfig": dbConf[db.StoredProcedureDb]})
