+++ metricsforwarder/cmd/metricsforwarder/main.go
@@ -41,8 +41,17 @@ func main() {
 	bindingDB := sqldb.CreateBindingDB(conf.Db[db.BindingDb], logger)
 	defer func() { _ = bindingDB.Close() }()
 
-	credentialProvider := cred_helper.CredentialsProvider(conf.CredHelperImpl, conf.StoredProcedureConfig, conf.Db, conf.CacheTTL, conf.CacheCleanupInterval, logger, policyDb)
-	defer func() { _ = credentialProvider.Close() }()
+	var credentialProvider cred_helper.Credentials
+	if conf.CredentialHelperConfig != nil {
+		credentialProvider = cred_helper.CredentialsProvider(
+			conf.CredentialHelperConfig,
+			conf.Db, conf.CacheTTL, conf.CacheCleanupInterval, logger, policyDb)
+	}
+	defer func() {
+		if credentialProvider != nil {
+			_ = credentialProvider.Close()
+		}
+	}()
 
 	httpStatusCollector := healthendpoint.NewHTTPStatusCollector("autoscaler", "metricsforwarder")
