+++ metricsforwarder/config/config.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"time"
 
-	"code.cloudfoundry.org/app-autoscaler/src/autoscaler/configutil"
 	"code.cloudfoundry.org/app-autoscaler/src/autoscaler/db"
 	"code.cloudfoundry.org/app-autoscaler/src/autoscaler/helpers"
 	"code.cloudfoundry.org/app-autoscaler/src/autoscaler/models"
@@ -34,74 +33,20 @@ type SyslogConfig struct {
 }
 
 type Config struct {
-	Logging               helpers.LoggingConfig         `yaml:"logging"`
-	Server                helpers.ServerConfig          `yaml:"server"`
-	LoggregatorConfig     LoggregatorConfig             `yaml:"loggregator"`
-	SyslogConfig          SyslogConfig                  `yaml:"syslog"`
-	Db                    map[string]db.DatabaseConfig  `yaml:"db"`
-	CacheTTL              time.Duration                 `yaml:"cache_ttl"`
-	CacheCleanupInterval  time.Duration                 `yaml:"cache_cleanup_interval"`
-	PolicyPollerInterval  time.Duration                 `yaml:"policy_poller_interval"`
-	Health                helpers.HealthConfig          `yaml:"health"`
-	RateLimit             models.RateLimitConfig        `yaml:"rate_limit"`
-	CredHelperImpl        string                        `yaml:"cred_helper_impl"`
-	StoredProcedureConfig *models.StoredProcedureConfig `yaml:"stored_procedure_binding_credential_config"`
-}
+	Logging              helpers.LoggingConfig
+	Server               helpers.ServerConfig
+	LoggregatorConfig    LoggregatorConfig
+	SyslogConfig         SyslogConfig
+	Db                   map[string]db.DatabaseConfig
+	CacheTTL             time.Duration
+	CacheCleanupInterval time.Duration
+	PolicyPollerInterval time.Duration
+	Health               helpers.HealthConfig
+	RateLimit            models.RateLimitConfig
 
-func LoadConfig(filepath string, vcapReader configutil.VCAPConfigurationReader) (*Config, error) {
-	conf := defaultConfig()
-
-	if err := helpers.LoadYamlFile(filepath, &conf); err != nil {
-		return nil, err
-	}
-
-	if err := loadVcapConfig(&conf, vcapReader); err != nil {
-		return nil, err
-	}
-
-	return &conf, nil
-}
-
-func defaultConfig() Config {
-	return Config{
-		Server:  helpers.ServerConfig{Port: 6110},
-		Logging: helpers.LoggingConfig{Level: "info"},
-		LoggregatorConfig: LoggregatorConfig{
-			MetronAddress: DefaultMetronAddress,
-		},
-		Health:               helpers.HealthConfig{ServerConfig: helpers.ServerConfig{Port: 8081}},
-		CacheTTL:             DefaultCacheTTL,
-		Db:                   make(map[string]db.DatabaseConfig),
-		CacheCleanupInterval: DefaultCacheCleanupInterval,
-		PolicyPollerInterval: DefaultPolicyPollerInterval,
-		RateLimit: models.RateLimitConfig{
-			MaxAmount:     DefaultMaxAmount,
-			ValidDuration: DefaultValidDuration,
-		},
-	}
-}
-
-func loadVcapConfig(conf *Config, vcapReader configutil.VCAPConfigurationReader) error {
-	if !vcapReader.IsRunningOnCF() {
-		return nil
-	}
-
-	conf.Server.Port = vcapReader.GetPort()
-	if err := configutil.LoadConfig(&conf, vcapReader, "metricsforwarder-config"); err != nil {
-		return err
-	}
-
-	if err := vcapReader.ConfigureDatabases(&conf.Db, conf.StoredProcedureConfig, conf.CredHelperImpl); err != nil {
-		return err
-	}
-
-	tls, err := vcapReader.MaterializeTLSConfigFromService("syslog-client")
-	if err != nil {
-		return err
-	}
-	conf.SyslogConfig.TLS = tls
-
-	return nil
+	// CredentialHelperConfig configures how credentials for "Basic Authentication" are managed.
+	// nil means no credential helper is configured (Basic Auth disabled).
+	CredentialHelperConfig models.BasicAuthHandlingImplConfig
 }
 
 func (c *Config) Validate() error {
@@ -175,8 +120,8 @@ func (c *Config) validateRateLimit() error {
 }
 
 func (c *Config) validateCredHelperImpl() error {
-	if c.CredHelperImpl == "" {
-		return errors.New("CredHelperImpl is empty")
+	if c.CredentialHelperConfig == nil {
+		return errors.New("CredentialHelperConfig is not configured")
 	}
 	return nil
 }
